move core oauth functionality out of the web module
