+++ cas-management-webapp-support/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     compile libraries.spring
     compile libraries.casclient
     runtime libraries.hibernate
+    runtime libraries.groovy
     compile project(':cas-server-core-services')
     compile project(':cas-server-core-authentication')
     compile project(':cas-server-core')
@@ -13,6 +14,6 @@ dependencies {
     compile project(':cas-server-core-logging')
     compile project(':cas-server-core-tickets')
     compile project(':cas-server-core-audit')
-    compile project(':cas-server-support-oauth')
+    compile project(':cas-server-support-oauth-core')
     compile libraries.javaximpl
 }
