+++ cas-management-webapp/src/main/webapp/WEB-INF/cas-management.properties
@@ -4,7 +4,7 @@ cas.prefix=${cas.host}
 cas.securityContext.casProcessingFilterEntryPoint.loginUrl=${cas.prefix}/login
 
 # Management
-cas-management.host=https://localhost:8443
+cas-management.host=https://mmoayyed.unicon.net:8443
 cas-management.prefix=${cas-management.host}/cas-management
 cas-management.securityContext.serviceProperties.service=${cas-management.prefix}/callback
