+++ cas-server-integration-mongo/build.gradle
@@ -7,7 +7,7 @@ dependencies {
     compile libraries.persondirectory
     runtime libraries.hibernate
     testCompile project(':cas-server-core-authentication')
-    testCompile project(':cas-server-support-oauth')
+    testCompile project(':cas-server-support-oauth-core')
     testCompile project(':cas-server-core-tickets')
     testCompile project(':cas-server-core-logout')
     testCompile project(path: ":cas-server-core-services", configuration: "tests")
