+++ cas-server-support-jpa-service-registry/build.gradle
@@ -9,7 +9,7 @@ dependencies {
     testCompile libraries.tests
     testCompile project(':cas-server-core-authentication')
     testCompile project(':cas-server-core-tickets')
-    testCompile project(':cas-server-support-oauth')
+    testCompile project(':cas-server-support-oauth-core')
     testCompile project(':cas-server-core-logout')
 }
