+++ cas-server-support-ldap-service-registry/build.gradle
@@ -6,7 +6,7 @@ dependencies {
     testCompile libraries.tests    
     testCompile project(path: ":cas-server-support-ldap", configuration: "tests")
     testCompile project(':cas-server-core-authentication')
-    testCompile project(':cas-server-support-oauth')
+    testCompile project(':cas-server-support-oauth-core')
     testCompile project(':cas-server-core-tickets')
     testCompile project(':cas-server-core-logout')
 }
