+++ cas-server-support-oauth-core/src/main/java/org/jasig/cas/support/oauth/services/OAuthRegisteredService.java
