+++ cas-server-support-oauth-core/src/test/java/org/jasig/cas/support/oauth/services/OAuthRegisteredServiceTests.java
