+++ cas-server-support-oauth/src/test/java/org/jasig/cas/support/oauth/OAuthTestSuite.java
