+++ settings.gradle
@@ -47,6 +47,7 @@ include ':cas-server-support-trusted'
 include ':cas-server-support-trusted-webflow'
 include ':cas-server-support-x509'
 include ':cas-server-support-x509-webflow'
+include ':cas-server-support-oauth-core'
 include ':cas-server-support-oauth'
 include ':cas-server-support-pac4j'
 include ':cas-server-support-pac4j-webflow'
